# BUG/PERF: Series(index=MultiIndex).rename losing EA dtypes