- [x] closes #21055
- [x] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [x] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [x] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [x] Added an entry in the latest `doc/source/whatsnew/v2.0.0.rst` file if fixing a bug or adding a new feature.


`Series(index=MultiIndex).rename` loses EA dtypes: 

```
import pandas as pd
import numpy as np

lev0 = pd.Index(np.arange(1000), dtype="Int64").astype("category")
lev1 = pd.Index(np.arange(1000), dtype="Int64")

mi = pd.MultiIndex.from_product([lev0, lev1], names=["A", "B"])
ser1 = pd.Series(1, index=mi)
ser2 = ser1.rename({10: 11}, level=1)

print(ser2.index.dtypes)
```

main:
```
A       int64
B       int64
dtype: object
```

PR:
```
A    category
B       Int64
dtype: object
```

Perf improves as well:

```
%timeit ser1.rename({10: 11}, level=1)

1.25 s ± 52.8 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)  -> main
350 ms ± 8.29 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)  -> PR
```